ref:不要なファイルの削除とRootコンポーネントの整理
- App.tsxとApp.cssを削除する
- ナビゲーションバーとRootコンポーネントを追加する
